Gameplay Updates
Added basic functionality for end point and scare point. End point will end the level, scare points will act as triggers for scare events. When the player enters a scare point, the scare point is destroyed and the player is given a value to it's fear level. Perhaps I'll adjust it to just deactivate until a set amount of time goes by BUT FOR DAY ONE we have the bare bones of a game loop!