Smol updates
Camera detects enemies in range so in the future the camera will swing around to look at the enemy at an offset and allow time for an animation sequence. Made some basic Main Menu and Credits screens, as well as included the bare scenes for the Carnival and Haunted house. The workshop will not be in the final build. Will continue work on this in awhile and get the camera done.